+++ CHANGELOG.md
@@ -1,3 +1,6 @@
-## 1.0.0
+## 1.0.1
 
-- Initial version, created by Stagehand
+- Connection with Google Tasks API (with authorization)
+- Displaying task lists
+- Adding new task to list
+- Displaying tasks in list